Latest master from upstream

+++ test/lock_test.py
@@ -1,30 +1,35 @@
-# Copyright (c) 2012 Spotify AB
+# -*- coding: utf-8 -*-
 #
-# Licensed under the Apache License, Version 2.0 (the "License"); you may not
-# use this file except in compliance with the License. You may obtain a copy of
-# the License at
+# Copyright 2012-2015 Spotify AB
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
 #
 # http://www.apache.org/licenses/LICENSE-2.0
 #
 # Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
-# WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
-# License for the specific language governing permissions and limitations under
-# the License.
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
 
-import unittest
+import os
+import subprocess
+import tempfile
+import mock
+from helpers import unittest
+
 import luigi
 import luigi.lock
 import luigi.notifications
-import tempfile
-import os
-import hashlib
-import subprocess
 
 luigi.notifications.DEBUG = True
 
 
 class TestCmd(unittest.TestCase):
+
     def test_getpcmd(self):
         p = subprocess.Popen(["sleep", "1"])
         self.assertTrue(
@@ -40,8 +45,15 @@ class LockTest(unittest.TestCase):
         self.pid, self.cmd, self.pid_file = luigi.lock.get_info(self.pid_dir)
 
     def tearDown(self):
-        os.remove(self.pid_file)
+        if os.path.exists(self.pid_file):
+            os.remove(self.pid_file)
         os.rmdir(self.pid_dir)
+
+    def test_get_info(self):
+        p = subprocess.Popen(["yes", "à我ф"], stdout=subprocess.PIPE)
+        pid, cmd, pid_file = luigi.lock.get_info(self.pid_dir, p.pid)
+        p.kill()
+        self.assertEqual(cmd, 'yes à我ф')
 
     def test_acquiring_free_lock(self):
         acquired = luigi.lock.acquire_for(self.pid_dir)
@@ -62,7 +74,7 @@ class LockTest(unittest.TestCase):
         self.assertTrue(acquired)
 
         s = os.stat(self.pid_file)
-        self.assertEqual(s.st_mode & 0777, 0777)
+        self.assertEqual(s.st_mode & 0o777, 0o777)
 
     def test_acquiring_lock_from_missing_process(self):
         fake_pid = 99999
@@ -73,4 +85,14 @@ class LockTest(unittest.TestCase):
         self.assertTrue(acquired)
 
         s = os.stat(self.pid_file)
-        self.assertEqual(s.st_mode & 0777, 0777)
+        self.assertEqual(s.st_mode & 0o777, 0o777)
+
+    @mock.patch('os.kill')
+    def test_take_lock_with_kill(self, kill_fn):
+        with open(self.pid_file, 'w') as f:
+            f.write('%d\n' % (self.pid,))
+
+        kill_signal = 77777
+        acquired = luigi.lock.acquire_for(self.pid_dir, kill_signal=kill_signal)
+        self.assertTrue(acquired)
+        kill_fn.assert_called_once_with(self.pid, kill_signal)
